requirements.txt
google-cloud-bigquery==2.3.1
Flask==1.1.2
flask-wtf
pandas-gbq
pandas
pyarrow
firebase-admin